First integration test passing
